feat: remove old breadcrumbs components and compact them into one file
